Create-JWT-service: move JwtEntity to entity package
